% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryoutube-package.R
\docType{package}
\name{ryoutube-package}
\alias{ryoutube}
\alias{ryoutube-package}
\title{ryoutube: Work with 'YouTube API'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provide function for get data from 'YouTube Data API' 
    <https://developers.google.com/youtube/v3/docs/>, 'YouTube Analytics API' 
    <https://developers.google.com/youtube/analytics/reference/> and 
    'YouTube Reporting API' <https://developers.google.com/youtube/reporting/v1/reports>.
}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

Other contributors:
\itemize{
  \item Netpeak [copyright holder]
}

}
