% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_auth.R
\name{ryt_deauth}
\alias{ryt_deauth}
\title{Suspend authorization}
\usage{
ryt_deauth()
}
\value{
only suspend authorization
}
\description{
Put ryoutube into a de-authorized state. Instead of sending a token,
ryoutube will send an API key. This can be used to access public
resources for which no Google sign-in is required. This is handy for using
ryoutube in a non-interactive setting to make requests that do not
require a token. It will prevent the attempt to obtain a token
interactively in the browser. The user can configure their own API key
via \code{\link[=ryt_auth_configure]{ryt_auth_configure()}} and retrieve that key via
\code{\link[=ryt_api_key]{ryt_api_key()}}.
In the absence of a user-configured key, a built-in default key is used.
}
\seealso{
Other auth functions: 
\code{\link{ryt_auth_configure}()},
\code{\link{ryt_auth}()}
}
\concept{auth functions}
