% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_playlists.R
\name{ryt_get_playlists}
\alias{ryt_get_playlists}
\title{Get playlist from 'YouTube API'}
\usage{
ryt_get_playlists(
  fields = c("contentDetails", "id", "localizations", "player", "snippet", "status")
)
}
\arguments{
\item{fields}{Fields of playist metadata, see \href{https://developers.google.com/youtube/v3/docs/playlists/list}{API documentation}.}
}
\value{
tibble with playlist metadata
}
\description{
Get playlist from 'YouTube API'
}
