% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_report.R
\name{ryt_get_report}
\alias{ryt_get_report}
\title{Load report data}
\usage{
ryt_get_report(download_url)
}
\arguments{
\item{download_url}{download URL, you can get it by \code{\link{ryt_reports_get_report_list}} or \code{\link{ryt_get_report_metadata}}}
}
\value{
tibble with report data
}
\description{
Load report data
}
\examples{
\dontrun{
 # auth
 ryt_auth('me@gmail.com')

 # get reporting data
 ## create job
 ryt_reports_create_job('channel_basic_a2')

 ## get job list
 jobs <- ryt_reports_get_job_list()

 ## get job report list
 reports <- ryt_reports_get_report_list(
   job_id = jobs$id[1],
   created_after = '2021-10-20T15:01:23.045678Z'
 )

 # get report data
 data <- ryt_get_report(
   download_url = reports$downloadUrl[1]
 )

 # delete job
 ryt_reports_delete_job(jobs$id[1])
 }
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reports}{Reporting API Documentation}

Other reporting api functions: 
\code{\link{ryt_get_report_metadata}()},
\code{\link{ryt_get_report_types}()},
\code{\link{ryt_reports_create_job}()},
\code{\link{ryt_reports_delete_job}()},
\code{\link{ryt_reports_get_job_list}()},
\code{\link{ryt_reports_get_report_list}()}
}
\concept{reporting api functions}
