% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_report.R
\name{ryt_get_report_metadata}
\alias{ryt_get_report_metadata}
\title{Retrieves the metadata for a specific report.}
\usage{
ryt_get_report_metadata(job_id, report_id)
}
\arguments{
\item{job_id}{job_id The ID that YouTube uses to uniquely identify the job for which reports are being listed. Use \code{\link{ryt_reports_get_job_list}}.}

\item{report_id}{The ID that YouTube uses to uniquely identify the report that is being retrieved. Use \code{\link{ryt_reports_get_report_list}}}
}
\value{
list with report metadata
}
\description{
Retrieves the metadata for a specific report.
}
\seealso{
Other reporting api functions: 
\code{\link{ryt_get_report_types}()},
\code{\link{ryt_get_report}()},
\code{\link{ryt_reports_create_job}()},
\code{\link{ryt_reports_delete_job}()},
\code{\link{ryt_reports_get_job_list}()},
\code{\link{ryt_reports_get_report_list}()}
}
\concept{reporting api functions}
