% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_report.R
\name{ryt_get_report_types}
\alias{ryt_get_report_types}
\title{Returns a list of report types that the channel or content owner can retrieve. Each item in the list contains an id property, which identifies the report's ID, and you need this value to schedule a reporting job.}
\usage{
ryt_get_report_types()
}
\value{
tibble with report types
}
\description{
Returns a list of report types that the channel or content owner can retrieve. Each item in the list contains an id property, which identifies the report's ID, and you need this value to schedule a reporting job.
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reference/rest/v1/reportTypes/list}{Reporting API Dicumentation}

Other reporting api functions: 
\code{\link{ryt_get_report_metadata}()},
\code{\link{ryt_get_report}()},
\code{\link{ryt_reports_create_job}()},
\code{\link{ryt_reports_delete_job}()},
\code{\link{ryt_reports_get_job_list}()},
\code{\link{ryt_reports_get_report_list}()}
}
\concept{reporting api functions}
