% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_auth.R
\name{ryt_has_token}
\alias{ryt_has_token}
\title{Is there a token on hand?}
\usage{
ryt_has_token()
}
\value{
Logical.
}
\description{
Reports whether ryoutube has stored a token, ready for use in downstream
requests.
}
\seealso{
Other low-level API functions: 
\code{\link{ryt_token}()}
}
\concept{low-level API functions}
