% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_report.R
\name{ryt_reports_create_job}
\alias{ryt_reports_create_job}
\title{Creates a reporting job.}
\usage{
ryt_reports_create_job(
  report_type = c("channel_annotations_a1", "channel_basic_a2", "channel_cards_a1",
    "channel_combined_a2", "channel_demographics_a1", "channel_device_os_a2",
    "channel_end_screens_a1", "channel_playback_location_a2", "channel_province_a2",
    "channel_sharing_service_a1", "channel_playback_location_a2", "channel_province_a2",
    "channel_sharing_service_a1", "channel_subtitles_a2", "channel_traffic_source_a2",
    "playlist_basic_a1", "playlist_device_os_a1", "playlist_playback_location_a1",
    "playlist_province_a1", "playlist_traffic_source_a1")
)
}
\arguments{
\item{report_type}{The type of report that the job creates. The property value corresponds to the id of a reportType as retrieved from the \code{\link{ryt_get_report_types}} function.}
}
\value{
No return value, called for side effects
}
\description{
By creating a reporting job, you are instructing YouTube to generate that report on a daily basis. The report is available within 24 hours of the time that the job is created.
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reference/rest/v1/jobs/create}{Reporting API Documentation}

Other reporting api functions: 
\code{\link{ryt_get_report_metadata}()},
\code{\link{ryt_get_report_types}()},
\code{\link{ryt_get_report}()},
\code{\link{ryt_reports_delete_job}()},
\code{\link{ryt_reports_get_job_list}()},
\code{\link{ryt_reports_get_report_list}()}
}
\concept{reporting api functions}
