% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_report.R
\name{ryt_reports_delete_job}
\alias{ryt_reports_delete_job}
\title{Deletes a reporting job.}
\usage{
ryt_reports_delete_job(job_id)
}
\arguments{
\item{job_id}{The ID that YouTube uses to uniquely identify the job that is being deleted. Use \code{\link{ryt_reports_get_job_list}}.}
}
\value{
No return value, called for side effects
}
\description{
Deletes a reporting job.
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reference/rest/v1/jobs/delete}{Reporting API Documentation}

Other reporting api functions: 
\code{\link{ryt_get_report_metadata}()},
\code{\link{ryt_get_report_types}()},
\code{\link{ryt_get_report}()},
\code{\link{ryt_reports_create_job}()},
\code{\link{ryt_reports_get_job_list}()},
\code{\link{ryt_reports_get_report_list}()}
}
\concept{reporting api functions}
