% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_report.R
\name{ryt_reports_get_job_list}
\alias{ryt_reports_get_job_list}
\title{Lists reporting jobs that have been scheduled for a channel or content owner.}
\usage{
ryt_reports_get_job_list()
}
\value{
tibble with jobs metadata
}
\description{
Each resource in the response contains an id property, which specifies the ID that YouTube uses to uniquely identify the job. You need that ID to retrieve the list of reports that have been generated for the job or to delete the job.
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reference/rest/v1/jobs/list}{Reporting API Documentation}

Other reporting api functions: 
\code{\link{ryt_get_report_metadata}()},
\code{\link{ryt_get_report_types}()},
\code{\link{ryt_get_report}()},
\code{\link{ryt_reports_create_job}()},
\code{\link{ryt_reports_delete_job}()},
\code{\link{ryt_reports_get_report_list}()}
}
\concept{reporting api functions}
