% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_report.R
\name{ryt_reports_get_report_list}
\alias{ryt_reports_get_report_list}
\title{Lists reports that have been generated for the specified reporting job.}
\usage{
ryt_reports_get_report_list(
  job_id,
  created_after = NULL,
  start_time_at_or_after = NULL,
  start_time_before = NULL
)
}
\arguments{
\item{job_id}{The ID that YouTube uses to uniquely identify the job for which reports are being listed.}

\item{created_after}{If specified, this parameter indicates that the API response should only contain reports created after the specified date and time, including new reports with backfilled data. Note that the value pertains to the time that the report is created and not the dates associated with the returned data. The value is a timestamp in RFC3339 UTC "Zulu" format, accurate to microseconds. Example: \code{"2015-10-02T15:01:23.045678Z"}.}

\item{start_time_at_or_after}{This parameter indicates that the API response should only contain reports if the earliest data in the report is on or after the specified date. Whereas the createdAfter parameter value pertains to the time the report was created, this date pertains to the data in the report. The value is a timestamp in RFC3339 UTC "Zulu" format, accurate to microseconds. Example: \code{"2015-10-02T15:01:23.045678Z"}.}

\item{start_time_before}{This parameter indicates that the API response should only contain reports if the earliest data in the report is before the specified date. Whereas the createdAfter parameter value pertains to the time the report was created, this date pertains to the data in the report. The value is a timestamp in RFC3339 UTC "Zulu" format, accurate to microseconds. Example: \code{"2015-10-02T15:01:23.045678Z"}.}
}
\value{
tibble with reports metadata
}
\description{
Lists reports that have been generated for the specified reporting job.
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reference/rest/v1/jobs.reports/list}{Reporting API Documentation}

Other reporting api functions: 
\code{\link{ryt_get_report_metadata}()},
\code{\link{ryt_get_report_types}()},
\code{\link{ryt_get_report}()},
\code{\link{ryt_reports_create_job}()},
\code{\link{ryt_reports_delete_job}()},
\code{\link{ryt_reports_get_job_list}()}
}
\concept{reporting api functions}
