% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_auth.R
\name{ryt_token}
\alias{ryt_token}
\title{Produce configured token}
\usage{
ryt_token()
}
\value{
A \code{request} object (an S3 class provided by \link[httr:httr-package]{httr}).
}
\description{
For internal use or for those programming around the YouTube API.
Returns a token pre-processed with \code{\link[httr:config]{httr::config()}}. Most users
do not need to handle tokens "by hand" or, even if they need some
control, \code{\link[=ryt_auth]{ryt_auth()}} is what they need. If there is no current
token, \code{\link[=ryt_auth]{ryt_auth()}} is called to either load from cache or
initiate OAuth2.0 flow.
If auth has been deactivated via \code{\link[=ryt_deauth]{ryt_deauth()}}, \code{ryt_token()}
returns \code{NULL}.
}
\seealso{
Other low-level API functions: 
\code{\link{ryt_has_token}()}
}
\concept{low-level API functions}
