% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryoutube-package.R
\docType{package}
\name{rytstat-package}
\alias{rytstat}
\alias{rytstat-package}
\title{rytstat: Work with 'YouTube API'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide function for get data from 'YouTube Data API' \url{https://developers.google.com/youtube/v3/docs/}, 'YouTube Analytics API' \url{https://developers.google.com/youtube/analytics/reference/} and 'YouTube Reporting API' \url{https://developers.google.com/youtube/reporting/v1/reports}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://selesnow.github.io/rytstat/docs/}
  \item Report bugs at \url{https://github.com/selesnow/rytstat/issues}
}

}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

Other contributors:
\itemize{
  \item Netpeak [copyright holder]
}

}
