% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_channel_activities.R
\name{ryt_get_channel_activities}
\alias{ryt_get_channel_activities}
\title{Returns a list of channel activity events}
\usage{
ryt_get_channel_activities(fields = c("contentDetails", "id", "snippet"))
}
\arguments{
\item{fields}{Fields of channel metadata, see \href{https://developers.google.com/youtube/v3/docs/activities/list}{API documentation}.}
}
\value{
tibble with channel activies
}
\description{
Returns a list of channel activity events
}
\examples{
\dontrun{

channel_activities <- ryt_get_channel_activities()

}
}
