% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_channels.R
\name{ryt_get_channels}
\alias{ryt_get_channels}
\title{Get channel info from 'YouTube API'}
\usage{
ryt_get_channels(
  fields = c("contentDetails", "id", "snippet", "statistics", "status", "topicDetails")
)
}
\arguments{
\item{fields}{Fields of channel metadata, see \href{https://developers.google.com/youtube/v3/docs/channels/list}{API documentation}.}
}
\value{
tibble with channel metadata
}
\description{
Get channel info from 'YouTube API'
}
\examples{
\dontrun{
channels <- ryt_get_channels()

}
}
