% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_playlist_items.R
\name{ryt_get_playlist_items}
\alias{ryt_get_playlist_items}
\title{Get playlist items data on 'YouTube'}
\usage{
ryt_get_playlist_items(
  playlist_id,
  fields = c("contentDetails", "id", "snippet", "status"),
  cl = NULL
)
}
\arguments{
\item{playlist_id}{Playlist ID.}

\item{fields}{Fields of video metadata, see \href{https://developers.google.com/youtube/v3/docs/playlistItems/list}{API documentation}.}

\item{cl}{A cluster object created by \code{\link{makeCluster}}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations (see Details on performance).}
}
\value{
tibble with playlist items details
}
\description{
Get playlist items data on 'YouTube'
}
