% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryt_get_video_details.R
\name{ryt_get_video_details}
\alias{ryt_get_video_details}
\title{Get detail data of your videos on 'YouTube'}
\usage{
ryt_get_video_details(
  video_id,
  fields = c("contentDetails", "fileDetails", "id", "liveStreamingDetails",
    "localizations", "player", "processingDetails", "recordingDetails", "snippet",
    "statistics", "status", "suggestions", "topicDetails"),
  cl = NULL
)
}
\arguments{
\item{video_id}{Video ID.}

\item{fields}{Fields of video metadata, see \href{https://developers.google.com/youtube/v3/docs/videos/list}{API documentation}.}

\item{cl}{A cluster object created by \code{\link{makeCluster}}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations (see Details on performance).}
}
\value{
tibble with video details
}
\description{
Get detail data of your videos on 'YouTube'
}
